+++ inc/jisonWrapper.ts
@@ -7,7 +7,7 @@ export class Parser {
 	parser: any;
 
 
-	constructor (grammar) {
+	constructor (grammar: string) {
 		// mute jison logs during grammar loading
 		const logs = [];
 		jison.print = log => logs.push(log);
@@ -22,8 +22,13 @@ export class Parser {
 	}
 
 
-	parse (source) {
+	parse (source: string): object {
 		//console.log("parse:", source.length);
 		return this.parser.parse(source);
+	}
+
+
+	generate (): string {
+		return this.parser.generate();
 	}
 };