+++ changelog.md
@@ -24,6 +24,8 @@
 
 *	2020.8.9
 
+	+	canvasEngraver.ts: packing images added.
+
 	+	tests\canvasEngraver.ts created.
 
 	+	sheet-live.vue: property of showPagesProgressively added.